# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/problems": {
      "post": {
        "operationId": "ProblemsController_createProblem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProblemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProblemDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Problems"
        ]
      },
      "get": {
        "operationId": "ProblemsController_getProblems",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BASICS",
                "CONDITIONS",
                "LOOPS",
                "STRINGS",
                "LISTS",
                "MATH",
                "ALGORITHMS"
              ]
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "EASY",
                "MEDIUM",
                "HARD"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProblemSummaryDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "total",
                        "page",
                        "limit",
                        "totalPages"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Problems"
        ]
      }
    },
...
    "/api/v1/problems/{id}": {
      "patch": {
        "operationId": "ProblemsController_updateProblem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProblemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProblemDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Problems"
        ]
      },
      "delete": {
        "operationId": "ProblemsController_deleteProblem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Problems"
        ]
      },
      "get": {
        "operationId": "ProblemsController_getProblemById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProblemDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Problems"
        ]
      }
    },
    "/api/v1/problems/{id}/test-inputs": {
      "get": {
        "operationId": "ProblemsController_getTestInputs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TestInputsDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Problems"
        ]
      }
    },
...
    "/api/v1/test-cases/problem/{problemId}": {
      "get": {
        "operationId": "TestCasesController_getTestCasesByProblem",
        "parameters": [
          {
            "name": "problemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TestCaseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "total",
                        "page",
                        "limit",
                        "totalPages"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "TestCases"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coding Arena",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProblemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "category": {
            "type": "string",
            "enum": [
              "BASICS",
              "CONDITIONS",
              "LOOPS",
              "STRINGS",
              "LISTS",
              "MATH",
              "ALGORITHMS"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "isBuiltIn": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "timeLimitMs",
          "category",
          "difficulty",
          "isBuiltIn",
          "description"
        ]
      },
      "CreateProblemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "category": {
            "type": "string",
            "enum": [
              "BASICS",
              "CONDITIONS",
              "LOOPS",
              "STRINGS",
              "LISTS",
              "MATH",
              "ALGORITHMS"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          }
        },
        "required": [
          "title",
          "description",
          "timeLimitMs",
          "isActive"
        ]
      },
...
      "UpdateProblemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "category": {
            "type": "string",
            "enum": [
              "BASICS",
              "CONDITIONS",
              "LOOPS",
              "STRINGS",
              "LISTS",
              "MATH",
              "ALGORITHMS"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          }
        }
      },
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "ProblemSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "category": {
            "type": "string",
            "enum": [
              "BASICS",
              "CONDITIONS",
              "LOOPS",
              "STRINGS",
              "LISTS",
              "MATH",
              "ALGORITHMS"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "isBuiltIn": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "timeLimitMs",
          "category",
          "difficulty",
          "isBuiltIn"
        ]
      },
      "TestInputsDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "inputs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "problemId",
          "timeLimitMs",
          "inputs"
        ]
      },
...
      "TestCaseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "problemId": {
            "type": "string"
          },
          "inputData": {
            "type": "string"
          },
          "expectedOutput": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "problemId",
          "inputData",
          "expectedOutput",
          "isHidden"
        ]
      },
...
    }
  }
}